% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{case_prop}
\alias{case_prop}
\title{Expected case carriage from an odds ratio}
\usage{
case_prop(or, carriage)
}
\arguments{
\item{or}{Odds ratio to detect.}

\item{carriage}{Control / population carriage frequency (exposure prevalence),
in (0, 1).}
}
\value{
The expected carrier frequency among cases.
}
\description{
Expected case carriage from an odds ratio
}
