% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{one_in}
\alias{one_in}
\title{Convert a "1 in N" frequency to a proportion}
\usage{
one_in(N)
}
\arguments{
\item{N}{The denominator of a "1 in N" rate.}
}
\value{
The proportion \code{1 / N}.
}
\description{
Convert a "1 in N" frequency to a proportion
}
\examples{
one_in(1000)
}
